Fix project version issue and handle case where udphCMake is submodule.

+++ udphProject.cmake
@@ -26,7 +26,7 @@ endfunction()
 macro(project_create proj_name)
 	if(GIT_PROJECT)
 		git_update_information(${proj_name})
-		project(${proj_name} VERSION "${GIT_VERSION}" LANGUAGES CXX)
+		project(${proj_name} VERSION "${${proj_name}_GIT_VERSION}" LANGUAGES CXX)
 	else()
 		project(${proj_name} LANGUAGES CXX)
 	endif()
@@ -45,24 +45,26 @@ endmacro()
 macro(finalize)
 	if(GIT_PROJECT AND GIT_CLEAN_SUBMODULES)
 		foreach(ITEM ${PROJECT_NAME}_GIT_SUBMODULES_STORED)
-			if(NOT ${ITEM} IN_LIST ${PROJECT_NAME}_GIT_SUBMODULES)
-				if(GIT_CLEAN_SUBMODULES_FORCE)
-					execute_process(COMMAND ${GIT_EXECUTABLE} submodule deinit --force "${ITEM}"
+			if(NOT "${ITEM}" STREQUAL "${CMAKE_UDPH_PATH}")
+				if(NOT ${ITEM} IN_LIST ${PROJECT_NAME}_GIT_SUBMODULES)
+					if(GIT_CLEAN_SUBMODULES_FORCE)
+						execute_process(COMMAND ${GIT_EXECUTABLE} submodule deinit --force "${ITEM}"
+										WORKING_DIRECTORY "${CMAKE_CURRENT_SOURCE_DIR}"
+	 									RESULT_VARIABLE COMMAND_RESULT)
+					else()
+						execute_process(COMMAND ${GIT_EXECUTABLE} submodule deinit "${ITEM}"
+										WORKING_DIRECTORY "${CMAKE_CURRENT_SOURCE_DIR}"
+	 									RESULT_VARIABLE COMMAND_RESULT)
+					endif()
+					if(NOT COMMAND_RESULT EQUAL "0")
+						message(FATAL_ERROR "Unable to deinit submodule ${ITEM}.")
+					endif()
+					execute_process(COMMAND ${GIT_EXECUTABLE} rm "${ITEM}"
 									WORKING_DIRECTORY "${CMAKE_CURRENT_SOURCE_DIR}"
 	 								RESULT_VARIABLE COMMAND_RESULT)
-				else()
-					execute_process(COMMAND ${GIT_EXECUTABLE} submodule deinit "${ITEM}"
-									WORKING_DIRECTORY "${CMAKE_CURRENT_SOURCE_DIR}"
-	 								RESULT_VARIABLE COMMAND_RESULT)
-				endif()
-				if(NOT COMMAND_RESULT EQUAL "0")
-					message(FATAL_ERROR "Unable to deinit submodule ${ITEM}.")
-				endif()
-				execute_process(COMMAND ${GIT_EXECUTABLE} rm "${ITEM}"
-								WORKING_DIRECTORY "${CMAKE_CURRENT_SOURCE_DIR}"
-	 							RESULT_VARIABLE COMMAND_RESULT)
-				if(NOT COMMAND_RESULT EQUAL "0")
-					message(FATAL_ERROR "Unable to clean submodule ${ITEM}.")
+					if(NOT COMMAND_RESULT EQUAL "0")
+						message(FATAL_ERROR "Unable to clean submodule ${ITEM}.")
+					endif()
 				endif()
 			endif()
 		endforeach()
